Added support of multiple network interfaces

+++ PKGBUILD
@@ -2,7 +2,7 @@
 
 pkgname=simple-rt
 pkgver=20170510.132_c270fc9
-pkgrel=1
+pkgrel=2
 url="https://github.com/vvviperrr/SimpleRT"
 pkgdesc="SimpleRT - Reverse Tethering utility for Android"
 arch=('x86' 'x86_64')
@@ -24,7 +24,9 @@ build() {
 	cd "$srcdir/simple-rt/simple-rt-cli"
 	make
 	export SIMPLERT_INTERFACE=$(ip link show | grep "state UP" | awk -F ': ' '{ print $2}')
-	cat "$srcdir/simple-rt@.service" | sed -e "s/%i/$SIMPLERT_INTERFACE/g" > "$srcdir/simple-rt.service"
+        if [ `echo "$SIMPLERT_INTERFACE" | wc -l` -eq 1 ]; then 
+                cat "$srcdir/simple-rt@.service" | sed -e "s/%i/$SIMPLERT_INTERFACE/g" > "$srcdir/simple-rt.service"
+        fi
 }
 
 package() {
@@ -34,5 +36,7 @@ package() {
 	install -m755 "$srcdir/simple-rt/simple-rt-cli/simple-rt" "$pkgdir/usr/lib/simple-rt/simple-rt"
 	install -m755 "$srcdir/simple-rt/simple-rt-cli/iface_up.sh" "$pkgdir/usr/lib/simple-rt/iface_up.sh"
 	install -DT -m0755 "$srcdir/simple-rt@.service" "$pkgdir/usr/lib/systemd/system/simple-rt@.service"
-	install -DT -m0755 "$srcdir/simple-rt.service" "$pkgdir/usr/lib/systemd/system/simple-rt.service"
+        if [ -e "$srcdir/simple-rt.service" ]; then 
+                install -DT -m0755 "$srcdir/simple-rt.service" "$pkgdir/usr/lib/systemd/system/simple-rt.service"
+        fi
 }
